requirements.txt
PyPDF2
faiss-cpu
streamlit
python-dotenv
langchain-groq
sentence-transformers
streamlit_option_menu 
langchain-huggingface
langchain-text-splitters
langChain
langchain_community
pypdf
docx2txt
pinecone-client==5.0.1
tiktoken
fastapi
uvicorn
langchain_groq
werkzeug
python-multipart


